add stop, sort podinfo, cleanup kill

+++ public/Start-WebServer.ps1
@@ -15,6 +15,9 @@ function Start-WebServer
         $msg = @{
             starting   = "Starting web server ..."
             started    = "Web server started on port {0}."
+            stopped    = "The web server was stopped."
+            killed     = "The pod was willed."
+            stressing  = "A stress test was started."
         }
     }
 
@@ -47,18 +50,21 @@ function Start-WebServer
                 if ( $EnableConsoleLogs ) { Write-Info -m $log -nl }
 
                 $stressAndRedirect = $false
+                $stop = $false
+                $kill = $false
+                $CpuThreads = $env:PSPOD_TEST_CpuThreads ??= 0
+                $MemThreads = $env:PSPOD_TEST_MemThreads ??= 0
 
                 switch ($uri)
                 {
-                    "GET /userlog" { $contentPath = $WS_USR_LOG_PATH; break }
-                    "GET /applog"  { $contentPath = $WS_APP_LOG_PATH; break }
-                    "GET /stop"    { $wsListener.Stop(); break wsListener }
                     "GET /stress10" {
                         Remove-Item -Path Env:\PSPOD_TEST_*
                         $env:PSPOD_TEST_WarmUpInterval = 0
                         $env:PSPOD_TEST_StressDuration = 10
                         $env:PSPOD_TEST_StressInterval = 10
                         $env:PSPOD_TEST_RestInterval   = 0
+                        $env:PSPOD_TEST_CpuThreads     = $CpuThreads
+                        $env:PSPOD_TEST_MemThreads     = $MemThreads
                         $stressAndRedirect             = $true
                         break
                     }
@@ -68,6 +74,8 @@ function Start-WebServer
                         $env:PSPOD_TEST_StressDuration = 240
                         $env:PSPOD_TEST_StressInterval = 10
                         $env:PSPOD_TEST_RestInterval   = 20
+                        $env:PSPOD_TEST_CpuThreads     = $CpuThreads
+                        $env:PSPOD_TEST_MemThreads     = $MemThreads
                         $stressAndRedirect             = $true
                         break
                     }
@@ -79,6 +87,8 @@ function Start-WebServer
                         $env:PSPOD_TEST_RestInterval        = 1
                         $env:PSPOD_TEST_MaxIntervalDuration = 60
                         $env:PSPOD_TEST_RandomizeIntervals  = "s,r"
+                        $env:PSPOD_TEST_CpuThreads          = $CpuThreads
+                        $env:PSPOD_TEST_MemThreads          = $MemThreads
                         $stressAndRedirect                  = $true
                         break
                     }
@@ -93,32 +103,52 @@ function Start-WebServer
                         $stressAndRedirect                  = $true
                         break
                     }
-                    "GET /kill" {
-                        Get-Process "pwsh" | Stop-Process
-                        break
-                    }
-                    default { $contentPath = $WS_APP_LOG_PATH; break }
+                    "GET /userlog" { $contentPath = $WS_USR_LOG_PATH; break }
+                    "GET /applog"  { $contentPath = $WS_APP_LOG_PATH; break }
+                    "GET /stop"    { $stop = $true; break }
+                    "GET /kill"    { $kill = $true; break }
+                    default        { $contentPath = $WS_APP_LOG_PATH; break }
                 }
 
                 if ( $stressAndRedirect ) {
-                    Start-Process -FilePath "pwsh" -ArgumentList ('-File', $DOCKER_PATH)
+                    Write-Info -ps -m $msg.stressing
+                    $stressProc = Start-Process -FilePath "pwsh" -ArgumentList ('-File', $DOCKER_PATH) -PassThru
                     $response.Redirect('/applog')
+                    $response.Close()
+                }
+                elseif ( $stop ) {
+                    $response.Redirect('/applog')
+                    $response.Close()
+                    start-sleep -seconds 3
+                    $wsListener.Stop()
+                    $wsListener.Close()
+                    Write-Info -ps -m $msg.stopped
+                }
+                elseif ( $kill ) {
+                    $response.Redirect('/applog')
+                    $response.Close()
+                    if ( $stressProc ) {
+                        Stop-Process $stressProc
+                        get-job | stop-job
+                        get-job | Remove-Job
+                    }
+                    Write-Info -ps -m $msg.killed
+                    start-sleep -seconds 3
+                    Get-Process "pwsh" | Stop-Process
                 }
                 else {
                     $ResponseText = $((Get-Content -path $WS_HEADER_PATH) -Join "`r`n") +
                                     $((Get-Content -path $contentPath)    -Join "`r`n") +
                                     $((Get-Content -path $WS_FOOTER_PATH) -Join "`r`n")
-
                     $buffer = [Text.Encoding]::UTF8.GetBytes($ResponseText)
-
                     $response.AddHeader("Last-Modified", [IO.File]::GetLastWriteTime($contentPath).ToString('r'))
-                    $response.AddHeader("Server", "psStress")
+                    $response.AddHeader("Server", "psPodTester")
                     $response.SendChunked = $FALSE
                     $response.ContentType = "text/HTML"
                     $response.ContentLength64 = $buffer.Length
                     $response.OutputStream.Write($buffer, 0, $buffer.Length)
+                    $response.Close()
                 }
-                $response.Close()
             }
         }
         finally
